+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All changes to this project will be documented in this file. This version
 format adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [x.x.x] - xxxx-xx-xx
+
+- feat: add support for entrypoint override in job
+
 ## [0.2.0] - 2025-10-25
 
 - fix: Prevent panic when the main branch of the repo is not equal to master.